requirements.txt
Django==4.2.13
psycopg2-binary
pandas
openpyxl
whitenoise
PyMuPDF
python-docx
requests
scikit-learn
xgboost
beautifulsoup4
opencv-python-headless
olefile
Pillow
markdown2
django-cors-headers
better-profanity
gunicorn
numpy<2
redis==4.6.0
django-redis
python-dateutil
django-allauth[socialaccount]

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
